Room rent

+++ src/pers/hy/public_rooms/serviceImpl/UserManageImpl.java
@@ -15,31 +15,61 @@ public class UserManageImpl implements UserManage{
 	}
 	
 	public String addUser(UserAddForm userAddForm){
-		if(userAddForm.getId().equals("")||userAddForm.getPassword().equals("")){
-			return "userAdd";
-		}else{
-			User user=userDao.addUser(userAddForm);
-			if(user==null){
-				ActionContext ctx=ActionContext.getContext();
-				ctx.put("exist","exist");
+		
+		String sessionResult=HttpSessionFactory.getHttpSession("type");	
+		if(sessionResult==null){
+			return "userManage";
+		}else if(sessionResult.equals("1")){
+			return "userManage";
+		}else if(sessionResult.equals("0")){
+			
+			if(userAddForm.getId().equals("")||userAddForm.getPassword().equals("")){
 				return "userAdd";
 			}else{
-				return "userManage";
+				User user=userDao.addUser(userAddForm);
+				if(user==null){
+					ActionContext ctx=ActionContext.getContext();
+					ctx.put("exist","exist");
+					return "userAdd";
+				}else{
+					return "userManage";
+				}	
 			}
-			
+		}else{
+			return "input";
 		}
 	}
 	
 	public String deleteUser(String[] userSelect){
+		
+		String sessionResult=HttpSessionFactory.getHttpSession("type");	
+		if(sessionResult==null){
+			return "userManage";
+		}else if(sessionResult.equals("1")){
+			return "userManage";
+		}else if(sessionResult.equals("0")){
 			userDao.deleteUser(userSelect);
-		    return "userManage";    
+		    return "userManage";  
+		}else{
+			return "input";
+		}	  
 	}
 
-	
 	public String getUserList(){
-		ActionContext ctx=ActionContext.getContext();
-		ctx.put("users", userDao.getUserList());
-		return "userManage";
-	}
+		String sessionResult=HttpSessionFactory.getHttpSession("type");	
+		if(sessionResult==null){
+			return "login";
+		}else if(sessionResult.equals("1")){
+			return "userIndex";
+		}else if(sessionResult.equals("0")){
+			ActionContext ctx=ActionContext.getContext();
+			ctx.put("users", userDao.getUserList());
+			return "userManage";
+		}else{
+			return "input";
+		}
+		
+		}
+	
 
 }
